Add fields to V4Environment
